refactor: change get accounts to command to test :hammer:
