+++ components/ledger/internal/services/command/get-ids-accounts.go
@@ -1,4 +1,4 @@
-package query
+package command
 
 import (
 	"context"
